move PatrolRewardPopup items to PatrolRewardModule
