+++ nekoyume/Assets/_Scripts/UI/Widget/Popup/PatrolRewardPopup.cs
@@ -4,7 +4,6 @@ using mixpanel;
 using Nekoyume.ApiClient;
 using Nekoyume.Game.Controller;
 using Nekoyume.L10n;
-using Nekoyume.UI.Model;
 using Nekoyume.UI.Module;
 using UnityEngine;
 using UnityEngine.UI;
@@ -19,10 +18,7 @@ namespace Nekoyume.UI
         [SerializeField] private PatrolRewardModule patrolRewardModule;
         [SerializeField] private ConditionalButton receiveButton;
 
-        public readonly PatrolReward PatrolReward = new();
-        private readonly Dictionary<PatrolRewardType, int> _rewards = new();
-
-        private bool _initialized;
+        public PatrolReward PatrolReward => patrolRewardModule.PatrolReward;
 
         protected override void Awake()
         {
@@ -47,11 +43,6 @@ namespace Nekoyume.UI
                 return;
             }
 
-            ShowAsync(ignoreShowAnimation);
-        }
-
-        private async void ShowAsync(bool ignoreShowAnimation = false)
-        {
             var clientInitialized = ApiClients.Instance.PatrolRewardServiceClient.IsInitialized;
             if (!clientInitialized)
             {
@@ -69,26 +60,8 @@ namespace Nekoyume.UI
             evt.AddCustomAttribute("patrol-time", PatrolReward.PatrolTime.Value.ToString());
             AirbridgeUnity.TrackEvent(evt);
 
-            if (!_initialized)
-            {
-                Init();
-            }
+            patrolRewardModule.SetData();
 
-            var avatarAddress = Game.Game.instance.States.CurrentAvatarState.address;
-            var agentAddress = Game.Game.instance.States.AgentState.address;
-            var level = Game.Game.instance.States.CurrentAvatarState.level;
-            if (!PatrolReward.Initialized)
-            {
-                await PatrolReward.InitializeInformation(avatarAddress.ToHex(),
-                    agentAddress.ToHex(), level);
-            }
-            else if (PatrolReward.NextLevel <= level)
-            {
-                await PatrolReward.LoadPolicyInfo(level);
-            }
-
-            SetIntervalText(PatrolReward.Interval);
-            OnChangeTime(PatrolReward.PatrolTime.Value);
             base.Show(ignoreShowAnimation);
         }
 
@@ -103,96 +76,14 @@ namespace Nekoyume.UI
             Close();
         }
 
-        // it must be called after PatrolReward.InitializeInformation (called avatar selected)
-        private void Init()
+        // subscribe from PatrolReward.PatrolTime, Claiming
+        private void SetReceiveButton(TimeSpan remainTime, bool claiming)
         {
-            PatrolReward.RewardModels
-                .Subscribe(OnChangeRewardModels)
-                .AddTo(gameObject);
-
-            PatrolReward.PatrolTime
-                .Where(_ => gameObject.activeSelf)
-                .Subscribe(OnChangeTime)
-                .AddTo(gameObject);
-
-            PatrolReward.Claiming.Where(claiming => claiming)
-                .Subscribe(_ => receiveButton.Interactable = false)
-                .AddTo(gameObject);
-
-            _initialized = true;
-        }
-
-        // rewardModels
-        private void OnChangeRewardModels(List<PatrolRewardModel> rewardModels)
-        {
-            SetRewards(rewardModels);
-        }
-
-        // time
-        private void OnChangeTime(TimeSpan patrolTime)
-        {
-            patrolTimeText.text =
-                L10nManager.Localize("UI_PATROL_TIME_FORMAT", GetTimeString(patrolTime));
-            patrolTimeGauge.fillAmount = (float)(patrolTime / PatrolReward.Interval);
-
-            var patrolTimeWithOutSeconds =
-                new TimeSpan(patrolTime.Ticks / TimeSpan.TicksPerMinute * TimeSpan.TicksPerMinute);
-            var remainTime = PatrolReward.Interval - patrolTimeWithOutSeconds;
             var canReceive = remainTime <= TimeSpan.Zero;
-
-            receiveButton.Interactable = canReceive && !PatrolReward.Claiming.Value;
+            receiveButton.Interactable = canReceive && !claiming;
             receiveButton.Text = canReceive
                 ? L10nManager.Localize("UI_GET_REWARD")
-                : L10nManager.Localize("UI_REMAINING_TIME", GetTimeString(remainTime));
-        }
-
-        private void SetIntervalText(TimeSpan interval)
-        {
-            gaugeUnitText1.text = GetTimeString(interval / 2);
-            gaugeUnitText2.text = GetTimeString(interval);
-        }
-
-        private static string GetTimeString(TimeSpan time)
-        {
-            var hourExist = time.TotalHours >= 1;
-            var minuteExist = time.Minutes >= 1;
-            var hourText = hourExist ? $"{(int)time.TotalHours}h " : string.Empty;
-            var minuteText = minuteExist || !hourExist ? $"{time.Minutes}m" : string.Empty;
-            return $"{hourText}{minuteText}";
-        }
-
-        private void SetRewards(List<PatrolRewardModel> rewardModels)
-        {
-            _rewards.Clear();
-
-            foreach (var reward in rewardModels)
-            {
-                var rewardType = GetRewardType(reward);
-                _rewards[rewardType] = reward.PerInterval;
-            }
-
-            patrolRewardModule.OnChangeRewards(_rewards);
-        }
-
-        private static PatrolRewardType GetRewardType(PatrolRewardModel reward)
-        {
-            if (reward.ItemId != null)
-            {
-                return reward.ItemId switch
-                {
-                    500000 => PatrolRewardType.ApStone,
-                    600201 => PatrolRewardType.GoldPowder,
-                    800201 => PatrolRewardType.SilverPowder,
-                    400000 => PatrolRewardType.Hourglass,
-                    _ => throw new ArgumentOutOfRangeException()
-                };
-            }
-
-            return reward.Currency switch
-            {
-                "CRYSTAL" => PatrolRewardType.Crystal,
-                _ => throw new ArgumentOutOfRangeException()
-            };
+                : L10nManager.Localize("UI_REMAINING_TIME", PatrolRewardModule.TimeSpanToString(remainTime));
         }
 
         // Invoke from TutorialController.PlayAction() by TutorialTargetType
